feat: Consolidate auth module into feature
