+++ app/build.gradle.kts
@@ -63,9 +63,7 @@ dependencies {
     implementation(projects.core.data)
     implementation(projects.core.database)
 
-    implementation(projects.auth.presentation)
-    implementation(projects.auth.domain)
-    implementation(projects.auth.data)
+    implementation(projects.feature.auth)
 
     implementation(projects.home.presentation)
     implementation(projects.home.domain)
