--- auth/domain/src/main/java/com/hyunjung/auth/domain/MyClass.kt
@@ -1,4 +1,0 @@
-package com.hyunjung.auth.domain
-
-class Test {
-}
