--- auth/presentation/src/main/java/com/hyunjung/auth/presentation/login/LogInAction.kt
@@ -1,7 +1,0 @@
-package com.hyunjung.auth.presentation.login
-
-interface LogInAction {
-    data object OnKakaoLogInClick : LogInAction
-    data object OnNaverLogInClick : LogInAction
-    data object OnGoogleLogInClick : LogInAction
-}
