+++ feature/auth/build.gradle.kts
@@ -1,0 +1,11 @@
+plugins {
+    alias(libs.plugins.hyunjung.cherrydan.android.feature.ui)
+}
+
+android {
+    namespace = "com.hyunjung.feature.auth"
+}
+
+dependencies {
+    implementation(projects.core.domain)
+}
