+++ feature/auth/src/main/AndroidManifest.xml
