+++ feature/auth/src/main/java/com/hyunjung/feature/auth/login/LogInAction.kt
@@ -1,0 +1,7 @@
+package com.hyunjung.feature.auth.login
+
+interface LogInAction {
+    data object OnKakaoLogInClick : LogInAction
+    data object OnNaverLogInClick : LogInAction
+    data object OnGoogleLogInClick : LogInAction
+}
