+++ settings.gradle.kts
@@ -24,9 +24,6 @@ enableFeaturePreview("TYPESAFE_PROJECT_ACCESSORS")
 rootProject.name = "Cherrydan"
 include(":app")
 include(":core")
-include(":auth:data")
-include(":auth:domain")
-include(":auth:presentation")
 include(":core:presentation:designsystem")
 include(":core:presentation:ui")
 include(":core:domain")
@@ -43,3 +40,4 @@ include(":search:presentation")
 include(":search:domain")
 include(":search:data")
 include(":core:model")
+include(":feature:auth")
